final commit

+++ db/migrate/20220705191725_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
       t.integer :user_id
-      t.integer :total
+      t.float :total
       t.string :status
 
       t.timestamps
